% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plickers.R
\docType{data}
\name{edu_ejemplo_datos_plickers}
\alias{edu_ejemplo_datos_plickers}
\title{Datos de ejemplo desde plickers}
\format{
Datos brutos, como se obtienen desde plickers, en esencia en csv, pero con las primeras filas con metadatos que son procesados
}
\source{
Obtenido desde una cuenta plickers siguiendo las instrucciones siguientes: https://help.plickers.com/hc/en-us/articles/4404834375707-How-to-export-student-results-in-a-datafile
}
\usage{
edu_ejemplo_datos_plickers
}
\description{
Conjunto de datos de ejemplo
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(edu_ejemplo_datos_plickers)
head(edu_ejemplo_datos_plickers)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
