% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\docType{data}
\name{edu_ejemplo_listado_estudiantes}
\alias{edu_ejemplo_listado_estudiantes}
\title{Datos de ejemplo de estudiantes}
\format{
csv
}
\usage{
edu_ejemplo_listado_estudiantes
}
\description{
Conjunto de datos de ejemplo
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(listado_estudiantes)
head(listado_estudaintes)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
