% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{edu_join_similar}
\alias{edu_join_similar}
\title{Fusionar columnas con datos afines entre si}
\usage{
edu_join_similar(tibble_procesar, nombre, matches_)
}
\arguments{
\item{tibble_procesar}{especifica la tibble en donde se encuentran las columnas que se desea unir}

\item{nombre}{especifica el nombre con el que se desee dejar las columnas que serán unidas}

\item{matches_}{especifica los nombres de las columnas que desean ser unidas, separadas por un signo |}
}
\description{
Función  que ofrece facilidades para unir datos comunes que provienen de distintas fuentes de datos
para que resulten tibbles con columnas unificadas.

Cuando se reunen datos por ejemplo de distintas hojas de excel, archivos csv, o algún otro,
puede suceder que en un origen la columna de nombre venga #' como 'nombre_alumno', en otra como 'alumno'
y en otra como 'alumnos'. Esta función ofrece facilidades para unificar bajo un mismo nombre.
}
\section{Datos en formato tidy}{

Funciones para reorganizar los datos.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Esta es la tibble original

edu_ejemplo_listado_estudiantes

# Esta es la tibble a la cual se renombraron las columnas nombre_completo, nombre y nombres como una sola: alumno

edu_join_similar(e_ejemplo_estudiantes,"alumno","nombre_completo|nombre|nombres")

edu_ejemplo_listado_estudiantes |> edu_join_similar("alumno","nombre_completo|nombre|nombres")
\dontshow{\}) # examplesIf}
}
\concept{pruebas}
