% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plickers.R, R/plickers_datos_ejemplo.R
\docType{package}
\name{EduToolsR-package}
\alias{EduToolsR}
\alias{EduToolsR-package}
\title{EduToolsR: Herramientas Didácticas y de Gestión para Docentes}
\description{
Herramientas que facilitan la evaluación educativa y la gestión administrativa.

Herramientas que facilitan la evaluación educativa y la gestión administrativa.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fjrojast/EduToolsR}
  \item Report bugs at \url{https://github.com/fjrojast/EduToolsR/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/fjrojast/EduToolsR}
  \item Report bugs at \url{https://github.com/fjrojast/EduToolsR/issues}
}

}
\author{
\strong{Maintainer}: Fernando Rojas \email{fjrojast@hotmail.com} (\href{https://orcid.org/0009-0003-2883-8809}{ORCID})

}
\keyword{internal}
