% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plickers_datos_ejemplo.R
\docType{data}
\name{plickers_datos_ejemplo}
\alias{plickers_datos_ejemplo}
\title{Datos de ejmplo desde plickers}
\format{
Datos brutos, como se obtienen desde plickers, en esencia en csv, pero con las primeras filas con metadatos que son procesados
}
\source{
Obtenido desde una cuenta plickers siguiendo las instrucciones siguientes: https://help.plickers.com/hc/en-us/articles/4404834375707-How-to-export-student-results-in-a-datafile
}
\usage{
plickers_datos_ejemplo
}
\description{
Conjunto de datos de ejemplo
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(plickers_datos_ejemplo)
head(plickers_datos_ejemplo)
plickers_datos_ejemplo \%>\% read_plickers()
read_plickers(plickers_datos_ejemplo)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
