% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plickers.R
\name{read_plickers}
\alias{read_plickers}
\title{Convierte la información exportada desde Plickers a un formato tidy}
\usage{
read_plickers(origin, fecha="2023-01-01")
}
\arguments{
\item{origin}{Selecciona el archivo CSV que proviene de plickers}

\item{fecha}{fecha formato aaaa-mm-dd, dd-mm-aaaa}
}
\value{
Devuelve una lista en formato tidy que facilite el análisis
}
\description{
Extrae la información que se obtiene en el csv que plickers exporta, para convertirla a una tibble que sigue las convenciones tidy-data
se proporciona un ejemplo extraído directamente de plickers para que puedan hacerse pruebas incluso antes de interactuar y generar evaluaciones directamente en plickers.
utilizar data("edu_ejemplo_datos_plickers") para poner el contenido del archivo csv en .globalenv.
utilizar posteriormente read_plickers(origin=edu_ejemplo_datos_plickers)
}
\section{Importar datos}{

Funciones para transformar datos provenientes de varios servicios de evaluación educativa
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Ejemplos de uso de la función
edu_ejemplo_datos_plickers |> read_plickers()
read_plickers(edu_ejemplo_datos_plickers)
\dontshow{\}) # examplesIf}
}
